Assert only common key fields during unmarshal
The open ID config implementation only support RSA keys, and only those
with an 'n' and 'e'. If those fields are missing, the unmarshaling will
fail. Then, the open ID config fails if any key that does not meet these
requirements comes in, also if there is a viable key.

Fail only for missing common fields as 'Kty' and 'Kid', and skip
other key types and RSA keys with missing or invalid 'n' and 'e' fields
after unmarshalling.