Share context and cancel signal in result/download
Getting results from storage often means starting a lot of concurrent
downloads. Any of these downloads could fail, and a single failure means
the process as a whole should be cancelled.

The azblob package has built-in support for this by taking a context [1]
parameter, so just sharing this context among the sibling goroutines
should accomplish this.

Testing really then just boils down to making sure the cancel() signal
is actually called, as exactly how and when tasks are cancelled is
controlled by azblob. Manually performed experiments have shown that
sibling jobs return _, err where err is "cancelled" for the http
requests.

This only addresses the goroutine cancellations, and is not at all
sensitive to *why* a transfer failed. In practice, it is likely to be a
timing issue where the result is not-yet available in storage. It's
unclear at this time if that should mean block-and-retry or failure, but
in either case the shared-context-with-cancellation-signal is a good
approach.

[1] https://golang.org/pkg/context/