Add Keyring for authing result requests
With queries being async, there has to be a mechanism in place that
makes sure results are not leaked, i.e. results are only available to
whoever requested the data.

The plan is to return a token which should be set in the authorization
header which provides (read) access to the process result and process
status. The creation, signing, and validation of these tokens is what
the Keyring provides.