Require oneseismic authorization token for /result
Ensure that only the user initiating a request can access the /result
resource, by rejecting any incoming requests that don't set the
x-oneseismic-authorization header with a valid and signed token. Any
request with a missing or invalid token will be aborted.

It's expected that the token will be moved to the Authorization header
once it no longer depends on blob storage.