Store query results in a redis database
Write and read (partial) results to a Redis instance, rather than to
blob storage.

Experiments indicate that reading fragments from the blob store is
reasonably quick, it introduces too much latency when writing results
back. The write operations seems fairly quick, but it takes too long
before the partial results can be fetched. Roughly, getting a 4M slice
takes ~3-6s, whereas the fetches write is commited at approx 50ms a
piece concurrently. These numbers are all on a vanilla 3-machine in
azure.

Cue redis, which has a lot of interesting properties more-or-less
tailored for this use case. Redis might in fact be used to implement a
lot more features, which will discussed when those features mature.

This patch replaces the concurrent download-to-channel with a single
redis MGET. There are features for concurrent download, but without
benchmarks and proper testing it is not obvious that it will be faster
than to just schedule N downloads on the same connection, and it is a
much easier implementation for now.

The result header ($pid-header.json) is still written to and read from
blob storage. This detail makes the Result.Get way longer than it needs
to be, but it was left alone for this patch in order to keep the
changeset small and focused.

Results are available in storage for 10 minutes, after which they will
be deleted by Redis.