Add hiredis as C++ dependency
Add hiredis [1] as a dependency. Hiredis seems like a reasonable
starting point for a redis client - easy to install and available, and
a quite straight forward C interface. At this point it is not quite
understood what's needed of a client, in particular w.r.t. recovery
should the connection fail for some time, so hiredis' simplicity should
make it an appropriate tool for getting that experience.

[1] https://redislabs.com/lp/hiredis/