Add abstraction for redis-based storage
Introduce some utility functions for the concept of working storage, a
volatile data store for data that's highly temporal and volatile, i.e.
(partial) results and process status.

This is really just a super thin layer over hiredis, which in time could
be moved to either a different technology altogether, or a 3rd party C++
redis library. For now, this should make for a reasonably simple way of
writing data to redis.

There's been some effort in trying to keep too much of redis' name and
behaviour out of the interface, and rather model the conceptual problem.
It is not yet certain that redis is here to stay for the purpose, which
should make moving to different storage tech easier.