Write result header to redis, not blob
Store and query the result header in Redis, not in blob. Measuring shows
that while writing the to blob is fairly fast, the object is not
sufficintly quickly available for GET - there's a fairly long [1] time
between when it's written until it is available for GET.

While it sounds fast, with the move to redis and dedicated nodes for
storing, full Volve slices are fetched in a matter of 2-300ms, which
means that the round-trip to blob is a significant barrier.

It doesn't make much sense to use the blob store for result headers,
when partial results are stored in redis. The partial results are stored
with expiration time, but there's no automatic cleanup of blobs [2], and
result headers are by their very nature temporary. There's no extra
security to be had by having to round-trip the blob store for the
request manifest. Moving to redis for the header makes the code cleaner
and faster.

[1] a rough estimate is 5-600ms at best
[2] while the blob store might support it, it's not implemented