Use err, not pollerr for redis error checks
A separate pollerr variable was necessary when err was already used for
a more specific type than the error interface. Now that the azure
download is facilities are replaced with redis get, the err variable is
now error, not dlerror.

The use of pollerr was actually a bug during development, where the
following if err != nil checked err, not pollerr. It was hard to notice
that it was err, not pollerr being checked (goes to show the danger of
repetition). Now that err is error, pollerr can go away.