Use Authorization header for auth in /result
The Result.Get feature used to rely on on-behalf tokens to fetch data
from Azure. That's only half the story for authentication though,
since the intention of the /result api is that only the one that made
the request should be able to query a specific process. For this, a
second type of auth was needed.

This was done by passing a token in the x-oneseismic-authorization
header - the Authorization header was already in use for Azure auth. Now
that Azure isn't queried in /result, the Authorization header is free
for authing the result.