Add /config endpoint
Add a /config enpoint which clients and users can GET to get a dict of
instance-specific configuration parameters for tokens and scopes. The
intention is to make login less painful, and make clients more
self-repairing and adaptable for new deployments.

The onseismic.login module is made aware of this, by making the
--client-id, --auth-server, and --scopes arguments optional, and by
adding a new positional parameter url. The url is the base-url for a
oneseismic installation, and will set the options as specified by the
running instance. The parameters can still be passed, and will override
whatever the server provides.