Add working_storage::get()
The get() function maps onto redis' GET command, but only accepts
strings as return. The implementation is more-or-less copied from put(),
and the goal is some working code to get unit- and integration tests up
and running. There are some strong opportunities for refactoring that
would kill the repetition.

Some thought has to go into the design of working_storage for it not to
just be a poorly designed general purpose redis client library, but
since it's a problem for later, the silly get() implementation should be
useful for getting tests up and exploring the problem more.