Find & connect to redis for integration tests
Look for redis on the host, or optionally connect to a running, public
redis instance, and use that for integration tests.

A part of the test suite has been broken ever since using blob store was
abandoned for storing process data, in favour of redis. These tests are
made working again, by connecting to a redis instance that lives for the
test session [1], using cmake's test fixtures [2]. With that came some
minor redesign of the makefiles to make this experience smoother -
integration and unit tests are two different ctest entries, but both run
the same binary.

For most use, if redis-server and redis-cli are available, there should
be no friction and building & running tests will be as before the redis
changes first came and broke the tests. Both the address and port number
can, however, be configured as escape hatches.

A significant amount of lines are added to the makefile, but a good part
is comments elaborating briefly on the considerations that went into it.
cmake && make && test should be enough for virtually all use.

[1] of course, when using a public redis the starting and stopping does
    not apply
[2] https://crascit.com/2016/10/18/test-fixtures-with-cmake-ctest/