Limit redis random-port to [10000, 60000)
Generating random 5-digit numbers means generating illegal port numbers
on the regular. Reserved ports [0, 1000] obviously won't be generated
since the string is 5-digit, but most systems have a valid port range of
[0, 65535].

Fix this by still generating a valid port number in two steps, where the
most-significant-digit is limited to [1, 5], and the remaining for
digits use the full [0, 9] range.