Block /query for permissions with manifest GET
Move the blocking read of the cube manifest to immediate /query call, so
bad permissions or cube-not-found become synchronous HTTP errors rather
than a status that has to be read on the result. This should make
end-user behaviour more comfortable and responsive by stopping and
reporting obvious errors and bad requests earlier.

Fetch the manifest in the /query implementation, and pass the manifest
as-is over ZMQ to the manifest-server [1]. It makes a lot of sense to do
this in go as it's essentially a blocking read + continue, which is the
type of work it excels at. The manifest read *must* be blocking for
security purposes, as it's otherwise easy to attack the system by
scheduling a bunch of bad reads that would have to be handled
downstream.

This will without doubt lead to slightly higher network load on good
requests, since the schedule-request messages now are manifest-sized
larger, but the service components should be on a much tighter network
loop, so this should in practice not mean much. There is an opportunity
to optimize this further by stripping down the manifest before it's
forwarded to job partitioning and scheduling.

This will add the round-trip to azure latency to /query, but should not
increase the response time for a *result*, as this was always a blocking
read.

This change decimates the manifest-task integration tests, since reads
from storage are no longer performed. The dependency on storage will be
removed at some point, but for this change it's plenty to replace the
storage dependency with one that raises errors if it is ever invoked.

[1] now awfully named, but this will change