Make /result auth into middleware
Writing the token validation and verification in the /result/:pid made
perfect sense to get it tested and up and running, but in experimenting
with a /result/:pid/status call it became pretty clear that all /result
endpoints would have to auth the same way, which is a perfect fit for
the middleware. A nice effect of this is Result.Get being more focused
on clear, since it no longer has to deal with auth nonsense.

Move the token handling code into the auth module, and hook it into the
/result/:pid call through gin's middleware.

In the same pass, the HTTP status codes were revised (they were pretty
much thrown in there in the first implementation with little thought or
reference) and should be more in line with what to expect from
authorization failures. A few more success and failure situations are
added to the test suite.