Move process header parsing out of Result.Get
Parsing input is pretty tedious in go, and not really an interesting
part of the Result.Get function. Additionally, it's quite a useful thing
for other features of /result, so it makes perfect sense for it not to
be an internal detail of Result.Get.

In addition to putting it its own function, the terminology has been
cleaned up slightly. The name "process header" is properly introduced,
deprecating the concept "Meta" (although it is still (partial) result
metadata).

The use of "meta" in the remaining function body is not change, to keep
this patch limited to just ripping out the parsing.