Add /result/:pid/status
Add an endpoint for querying the status of a result. This allows clients
to check if the result is ready for fetching (and would surely return
200/OK), rather than requesting /result/:pid and waiting for a timeout.

This is just a rough implementation, with many opportunities for tests
and refactoring, but should be useful for getting some experience with
the design.