Add entry endpoint /query/:guid and cube.shape
Add a cube entry point that describes the available features for a cube,
and provide some basic metadata (for now, just the dimensions). The
dimension is re-introduced in Python through the more general cube.shape
attribute.

The implementation of slice.Entry is pretty rough - it does not really
belong on the Slice struct and is mostly copy-paste of the
fetch-and-parse logic from Slice.Get. It will certainly be refactored
quite soon, but is written as-is to highlight what exactly is (typical)
common code. Please note the absence of tests is with this in mind -
tests should be introduced with said refactoring.