Add test for azure-derived errors getting manifest
Add a test, and more importantly a tiny framework supporting it, for
verifying that abortOnManifestError() sets the right error code, writes
the correct messge, and properly aborts the request.

While the test itself is pretty trivial and unnecessary, this is only
because there is no handling of different types of errors. Handling
azure-originating errors is difficult without recording the behaviour in
tests, leading to a chicken-and-egg problem. This is an egg.