Set middleware on groups
All endpoints within a group will tend to rely on the same middleware.
Connecting the middleware on the group allows for a much less repetitive
and error prone list of middlewares on the GET() calls, and makes the
declaration much cleaner.

This is a nice preparation for when the endpoints will be split into
multiple processes.