Generate pid in middleware
Explicitly generating pid is fairly tedious, although it is about as
much repetition as looking up the pid on the context object.
Conceptually it is a much cleaner thing to shove into middleware than
having as an explicit operation in every handler.

pid is generated for tying requests and data to a user query for data,
but is also used in endpoints that only gets metadata. The driving
decision is to have more tracability and better logs, and to give end
users a way to point operators the right way when something goes wrong.
The name "pid" is in other words overloaded to work as the (maybe more
common) request-id.

The /result family does not generate a pid, since the pid is a parameter
to the request.

The overhead of using middleware for this is not measured, but it
*should* be reasonably insignificant.