Add util.GetManifest() fetch-and-parse function
Introduce the GetManifest() function that will fetch and parse, or
appropriately abort the request. I considered putting this into
middleware, but desided against it for now as there are enough
undocumented parameters being passed that way already, which makes it
difficult to get an overview of what code is run in each handler. It is
quite possible that future changes would tip the scales and move
GetManifest() into middleware.

All in all the goal is to make the handler functions more focused on
the interesting work, and not having too much noise from grabbing
dependencies and tedious error checking.