Add load balancing message broker
Message broker that will distribute tasks to workers as they become
available.

The advantage of this approach over PUSH-PULL is that jobs are sent to
idle workers, whereas PUSH-PULL distributes round robin. A disadvantage
is that the queue becomes a single point of failure, and a single point
for all messages to pass through, while PUSH-PULL is N-to-N.