Use requests.session to set headers
Set headers used in every request on a session object, rather than
looking it up from the cache every time. This is a small change in code,
but carries some deeper implications:

1. It is assumed that the authorization header will be unchanged
   throughout the lifetime of the client. Long-lived clients could bring
   all kinds of problems, so this is a feature.
2. It's a double-down on using the requests library to perform http
   requests, since more state is managed by it.

Doubling down on requests means porting to a different lib for http
requests is more work, since more assumptions and state is handed over
to it. The benefit is not having to handle so much state in oneseismic,
simply and needing less code (in particular for request headers).

Fewer requests parameters are set every call site, which trades
immediate and local code understanding for the global correctness,
performance, and maintainer burden.