Turn non-200 responses into exceptions
Raise exceptions when receiving non-200 responses. This makes it
impossible to silently continue due to forgetting to check the http
status code, and makes checking and handling http status derived
failures more obvious.

The requests library includes the response objects in the exception, so
there is no loss of power when handling errors. The response body is
still available.