Add object for representing on-going request
With the move to an asynchronous API design, the need arises for a way
to schedule server-side jobs without blocking python programs, and a
need to monitor the progress and status more ergonomically than
hand-crafting requests.

Introduce the process class. This is the low-level facilities for more
comfortably map between processes happening server-side, and the
high-level APIs intended for users. The implementation as it stands is a
little rough, but provides a more nautural place for the HTTP related
code.