Add request.Sesssion.get automation
Introduce the http_session child class for some automation and a central
point for oneseismic aware http functionality.

The http_session is based on the request.Session type to provide
connection keep-alive and persisting headers on a root "prototype
request". This makes working with token-based authentication more
ergonomic, since doesn't have to be stored in this layer, but handed
over to request.

Some minor renaming follows.