Add ijk property
Add the ijk property for clients to query the inline/crossline/depths,
i.e. the addressable samples. It shold hold that query/slice/<dim>/<no>
should be a valid request when <no> = ijk[<dim>][k]. The property relies
on a slightly beefier response from query/<guid>.

The name ijk is bad, but until some concept clears up this is the best I
could think of. I expect it to change at some point in the (near)
future.