Move ls() out of client.client
Move the ls() function out of client, and take only a session as
parameter.

This is a move towards re-designing (or killing) the client class, and
have functionality rely on a narrower dependency. The goal is to make
bits and pieces more composable, and more easily testable. The problem
with the client god object is that *some* assumptions are quite
desirable for high-level applications, such as always reading the
cache/config file, but they make testing and just-in-time configuration
difficult.

The client class was on its way to becoming a god object, which it
should not be.