Redesign scan and upload programs
This is one of those patches that ends up touching something
fundamentally wrong, which warrants a re-design and changes cascade. The
ingestion story has been poor for a long time, and with this patch is
finally reviewed.

Synopsis
--------
+ scan no longer reject unsorted files
+ scan walks the entire file and stores all in- and crosslines
+ scan tolerates missing in/crossline pairs
+ upload no longer cleverly scans multiple lines at once, works on
  trace-by-trace
+ upload tolerates files with holes, files that taper off
+ scan and upload can both read and write from blob store

Notes
-----
A lot of new data has been thrown at oneseismic, which highlighteded
some bad assumptions in the scan and upload producedure. In the wild, it
is too common to have surveys with holes on the edges, non-square areas
and missing lines, so the strict input-SEG-Y-must-be-regular assumption
is just too much of a burden.

The upload program was designed to be faster by leveraging the
assumption that files had no holes and did not taper off, but it turns
out that in practice it is not so worse to walk the full file. In some
pathological cases this leads to higher memory usage, but it can upload
files with traces all scrambled, which is covered in a test.

A new module 'internal' has been introduced, with tooling useful for
writing programs, but not polished or general enough to be their own
packages. This includes new I/O facilities for reading and writing
directly to/from the blob store, and some argument parsing.

In this iteration of the upload program, files with missing traces are
explicitly padded with zeros. This may change in the future if the
server learns how to handle subvolumes without data.

Examples
--------
Some examples to illustrate the use of the new programs.

Scan a file in azure with connection string auth:
    oneseimic scan https://acc.blob.net/volumes/vol1.sgy
                    --input-connection-string <connection-string>

Scan a file in azure with url-encoded auth:
    oneseimic scan https://acc.blob.net/volumes/vol1.sgy?<sas-token>

Scan a file in azure with explicit credentials:
    oneseimic scan https://acc.blob.net/volumes/vol1.sgy
                    --input-credentials <credentials>

Scan a file with explicitly chosen auth method:
    oneseimic scan https://acc.blob.net/volumes/vol1.sgy?<sas-token>
                    --input-connection-string <connection-string>
                    --input-credentials <credentials>
                    --input-auth-method connection-string

Upload a file from the local file system:
    oneseismic upload vol-scan.json /volumes/vol1 https://acc.blob.net/
                    --output-connection-string <connection-string>

Upload a file from from a blob, using a shared access signature:
    oneseismic upload vol-scan.json
                    https://inputacc.blob.net/vol1?<sas-token>
                    https://outputacc.blob.net/
                    --output-credentials <credentials>