Offload scan.scan/segment logic to scanner object
Offload the logic of scanning headers and building a report to a scanner
object, which is responsible for maintaining and index, maintain
invariants, and create a report.

This change in isolation introduce real change in behaviour, but paves
the way for different results when scanning headers, which is what quite
a few useful tasks on SEG-Y files boil down to.

This has the benefit of making the loop-over-headers code simpler, since
the interrogation and updating of state is handled by a scanner/action
object.