Port fragment/fetch server to go
Most of the fragment-fetch server is maintaining download https
sessions, with the goal of getting the fragment set *as fast as
possible* from the blob store.

To understand the motivation of the port, some historical context is
useful. When the the core/server-fragment was first written, there was
no operational services and no good idea of how fast the blob store could
deliver data. It was assumed that there would be a significant need to
turn the knobs of the http library in order to get good concurrent
performance. Some of it is true - the CPU load when fetching is pretty
much zero, and multiple concurrent downloads are necessary to even
remotly saturate the link [1]. Furthermore, the backing library that
extracted data from fragments was reasonably well tested, and also
written in C++, so writing the server in C++ meant integrating this code
was easy.

Now that oneseismic has matured, it seems that in order for downloading
sets of fragments to be fast, what's necessary for performance is quick
scheduling and concurrent connections. This is pretty much a description
of go's power, and writing sophisticated scheduling is a pretty awful
job that is in no way specific to oneseismic.

The cmd/fetch program implements a server that operates identically to
the C++ server-fragment, but in a lot less code and with some nice
concurrency and scheduling improvements to boot.

The implementation uses the C++ library to extract data from fragments,
which means go is really only responsible for managing http sessions,
I/O, and buffers.

Some rough measurements indicates that the new server is slightly faster
(~5-10%) , and I expect that with more tuning of code and parameters [2]
the new implementation will become significantly faster. Regardless of
speed, the code itself is far more managable, as it off-loads most of
the difficult scheduling and signalling work to the go runtime.

[1] which, at this point, I don't think we're able to
[2] concurrent connections in and connection re-use are particularly
    interesting