Merge Dockerfiles and change fetch program
The Dockerfiles were long overdue for some work. Now that a go program
depends on the C++, the build becomes more involved, and some
long-standing shortcomings are addressed.

Briefly, the dockerfiles are merged into a single multistage docker
file, and all executables are bundled together in a smaller release
image. This means that oneseismic.azurecr.io/base is the only image, and
/api is dead. The docker-compose is changed to reflect this, and the new
oneseismic-fetch program replaces the oneseismic-fragment program as the
fetch service.

The resulting image is slightly larger than the individual images before
(115M -> 140M), but once dependencies from now-deprecated programs are
removed the image size is expected to shrink.

The reviewed use of multistage builds means that clean builds should be
faster, and rebuilds when making changes in multiple languages should be
faster and use less space, since more stages are reusable.