Remove dead core/python library
This library has been dead for a while (it never really came online),
but carries some weight. Remove it outright, and if oneseismic should
pull in a direction where it becomes useful again it is probably a
better idea to start fresh than it is to twist this code into something
else.