Update go packages
Update packages to later versions and fix the few API breaks that
occured.

The most "controversial" thing that isn't just go tooling noise is

-       "github.com/dgrijalva/jwt-go"
+       "github.com/form3tech-oss/jwt-go"

which is done because the auth0/jwt-middleware imports the fork, which
made types incompatible. Both projects should be monitored in case
there's a schism and to figure out which project to build on.

This patch adds a bunch of packages that were imported, but never
checked into the go.mod file.