Remove dead code from /result
The /result endpoint has seen a lot of development, and a lot of cruft
has been sitting around for a while. Remove the unused functions and
their corresponding tests.

While the file is much smaller now, it became pretty clear that the
code supporting the /result endpoint is not really covered by tests.
There are many reasons (some good) for this, but it makes it stand out
as an opportunity for some improvement.