Support blob-specific SAS auth
The SAS authorization support in blobfs was poor, and did not work for
signatures that give access to a single blob. Single-blob access for
reading is a particularly common use case when ingesting.

The issue roots in the BlobServiceClient not being constructable from a
blob-specific SAS, as it provides an interface for the concept of the
blob storage account. Single-blob SAS can only be used to directly
construct a BlobClient, which previously was obtained by going
account.container.blob from BlobServiceClient.

To maintain the high-level filesystem metaphor, cd() and open() is still
supported and uses relative paths, but with some improved exceptions.