Dispatch tasks on redis stream
This is a first step in a re-design of the task distribution to use
redis streams rather than ZMQ.

Context
-------
For a long time, there has been an issue with the messages queues - they
would silently stall after a long period of inactivity, without any
party recognising it. The culprit is the TCP connection being dropped on
either side of the API <-> scheduler <-> worker boundary, and neither
application is aware.

The situation improved when the scheduler got its pull-based
ready-for-work design, but the connection between the API servers and
the scheduler would still break. This manifested as jobs being "passed
on" from the API server, but the scheduler never receiving the job. This
was mitigated by setting TCP-keepalive and ZMQ heartbeating, but it is
not sufficient for a robust design.

This is expected behaviour with ZMQ, and makes some sense - even without
TCP connections sliently being dropped, the applications need to be
aware of the state of the (immediate) peers for all sorts of robustness
and scaling reasons, especially with workers coming-and-going.

Leading up to this, I have done several experiments with keep-alive and
heartbeating between nodes. The implementation does not end up being so
bad, but naturally handling this robustly means a non-trivial amount of
code. Moving beyond the problem of connections silently dropping,
there's the issue of dynamic scaling, in particular of workers.

Enter redis
-----------
After some exploration of the greater issue, redis pops up again. Redis
comes with built-in support for fan-out work scheduling through its
stream and consumer groups features, which is exactly what the
scheduler/oneseismic-manifest program implemented.

This comes with a long list of benefits:

  - redis can be in charge of scaling, with replicas and what-not
  - the connection keep-alive problem is handed off to redis
  - new clients can easily join the worker pool with a single redis
    connect()
  - the stream can be audited with generic tools
  - the custom scheduler program stops existing

This means that the task queue is managed by redis, and the fetch client
is implemented with a simple redis blocking read.

The implementation
------------------
This patch moves the request -> task-set mapping to the API node, and
re-wires the fetch program to pull from redis rather than ZMQ,
eliminating the server-manifest altogether. No protocol or format
changes are introduced, only the transport has changed.

The creation of chunks-of-work objects is done in C++ (as it relies on
deep understanding of the geometry), but invoked from Go. This builds on
the pattern established by fetch, where Go is responsible for the
main-loop and I/O, and C++ understands the geometry and provides the
data.

The Go programs effectively run the C++ code as separate programs and
parse the output, but now that Go directly invoke C++ then there are
good opportunities for refactoring. No refactoring is done in this
patch, in order to make reverts and merges easier.