Delete jobs on receive; limit queue growth
The redis stream (XADD/XREAD/XREADGROUP) has no built-in expire or know
the read-and-forget semantics of oneseismic, so the strategy for
cleaning up processed items must be implemented. Without a concept of a
master node or more communication, the easiest system is for individual
nodes to delete jobs on read.

This is somewhat inefficient as there is now significantly more requests
going into redis. While the (expensive) delete actually happens at
redis' discretion, it will always be worse than more sweeping deletes.

This is the quick-and-easy solution for managing the job queue size, and
while it might not scale well enough™ it should be plenty for the
forseeable future.

Some other strategies and their issues:

* Using XADD with MAXLEN
        Effectively bound the queue at N elements. This N would have to
        be very high, since common jobs could easily spawn a few
        thousand requests, and you could still lose jobs in huge bursts
        of work.
* Explicit pruning with XTRIM MINID
        XTRIM MINID was introduced in redis 6.2, which is currently in
        pre-release. XINFO can ask the consumer group what ID was the
        last consumed, and since the queue is consumed linearly this ID
        can be used for deleting everything now-read.
* Periodic garbage collection
        This is an interesting idea because a periodic garbage collector
        can be put off in times of high load, and can also clean up dead
        consumers. It would require some cron-like facility or even more
        dynamic scheduling of gc runs, and some heuristic on how to
        prune. This could be easy, and it could be hard, and regardless
        it's significantly more work to get right than XDEL for every
        read.

This is an optimisation to bound memory use, and does not contribute to
correctness.