Fail on XGROUP CREATE hard errors
The XGROUP CREATE function can fail with hard errors (typically network
issues) or soft errors (group already exists). The latter is to be
expected - in fact, it should fail a lot more than it suceeds as only
the first worker to report in gets to create the group.

The intention is to make this reasonably clear from docs and control
flow.