Build go-binaries with single build command
Build the go cmd/binaries in a single command rather than a WORKDIR +
RUN build for every binary. This makes the image include new executables
without having to add a separate build command [1], and makes docker
generate fewer intermediate images. The caching of intermediate layers
for go binaries seems to work poorly anyway, so this is not a big loss.

[1] copying into the deploy image still requries an explicit copy