Reboot readme
We've been without a readme for a while, with focus going elsewhere, and
with the approach being reviewed. It is time to re-introduce it, as a
sort-of landing page for the peope curious about oneseismic.

This is an early draft with empty sections and it is sparse on
information, but it should serve well enough as a place to write as we
go. It will certainly improve quickly from here.