Explicitly request tokens, not though middleware
Rather than implicitly getting tokens through the middleware, make code
explicitly request it. This comes with a slightly higher code tax in
some cases, but opens up for new and interesting opportunities.

The middleware is great for a lot of things, but in the case of
acquiring on-behalf-of tokens it became too much of a burden. In terms
of code volume we're not much worse off, since we still needed to fetch
the value from the context - now we need to ask the token fetcher
directly.

The Tokens concept is as close to a global as you can get, but it is
implemented through struct-members regardless to help testing.